move DeviceScreenSize to new file
